quick pick - add quick pick service to platform and register for standalone editor
